+++ src/components/CategoryFilter.js
@@ -1,20 +1,18 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { cat } from './BookForm';
+import React from "react";
+import PropTypes from "prop-types";
+import { cat } from "../containers/BookForm";
 
-const CategoryFilter = props => {
+const CategoryFilter = (props) => {
   const { onChange } = props;
-  const handleChange = event => {
+  const handleChange = (event) => {
     const { value } = event.target;
     onChange(value);
   };
   return (
     <select name="category" onChange={handleChange}>
       <option>All</option>
-      {cat.map(category => (
-        <option key={category}>
-          {category}
-        </option>
+      {cat.map((category) => (
+        <option key={category}>{category}</option>
       ))}
     </select>
   );
